refactor: :recycle: split telemetry into own file
