+++ src/lib.rs
@@ -1,40 +1,4 @@
 pub mod configuration;
 pub mod routes;
 pub mod startup;
-// use actix_web::dev::Server;
-// use actix_web::{web, App, HttpRequest, HttpResponse, HttpServer, Responder};
-// use std::net::TcpListener;
-
-// async fn greet(req: HttpRequest) -> impl Responder {
-//     let name = req.match_info().get("name").unwrap_or("World");
-//     format!("Hello {}!", &name)
-// }
-
-// async fn health_check() -> impl Responder {
-//     HttpResponse::Ok().finish()
-// }
-
-// #[allow(dead_code)]
-// #[derive(serde::Deserialize)]
-// struct FormData {
-//     email: String,
-//     name: String,
-// }
-
-// async fn subscribe(_form: web::Form<FormData>) -> HttpResponse {
-//     HttpResponse::Ok().finish()
-// }
-
-// pub fn run(listener: TcpListener) -> Result<Server, std::io::Error> {
-//     let server = HttpServer::new(|| {
-//         App::new()
-//             .route("/health_check", web::get().to(health_check))
-//             .route("/subscriptions", web::post().to(subscribe))
-//             .route("/", web::get().to(greet))
-//             .route("/{name}", web::get().to(greet))
-//     })
-//     .listen(listener)?
-//     .run();
-
-//     Ok(server)
-// }
+pub mod telemetry;
